Modify program to use the new PitchSelector class

+++ Shift.cs
@@ -6,35 +6,35 @@ using System.Threading.Tasks;
 
 namespace Cnthesizer
 {
-	public delegate float Shift(float frequency);
+	public delegate double Shift(double frequency);
 
 	public static class Shifts
 	{
-		public static float OctaveUp(float frequency) => frequency * 2;
-		public static float MajSeventhUp(float frequency) => frequency / 8 * 15;
-		public static float MinSeventhUp(float frequency) => frequency / 9 * 16;
-		public static float MajSixthUp(float frequency) => frequency / 3 * 5;
-		public static float MinSixthUp(float frequency) => frequency / 5 * 8;
-		public static float PerfectFifthUp(float frequency) => frequency / 2 * 3;
-		public static float TritoneUp(float frequency) => frequency / 32 * 45;
-		public static float PerfectFourthUp(float frequency) => frequency / 3 * 4;
-		public static float MajThirdUp(float frequency) => frequency / 4 * 5;
-		public static float MinThirdUp(float frequency) => frequency / 5 * 6;
-		public static float MajSecondUp(float frequency) => frequency / 8 * 9;
-		public static float MinSecondUp(float frequency) => frequency / 15 * 16;
-		public static float Unison(float frequency) => frequency;
-		public static float MinSecondDown(float frequency) => frequency / 16 * 15;
-		public static float MajSecondDown(float frequency) => frequency / 9 * 8;
-		public static float MinThirdDown(float frequency) => frequency / 6 * 5;
-		public static float MajThirdDown(float frequency) => frequency / 5 * 4;
-		public static float PerfectFourthDown(float frequency) => frequency / 4 * 3;
-		public static float TritoneDown(float frequency) => frequency / 45 * 32;
-		public static float PerfectFifthDown(float frequency) => frequency / 3 * 2;
-		public static float MinSixthDown(float frequency) => frequency / 8 * 5;
-		public static float MajSixthDown(float frequency) => frequency / 5 * 3;
-		public static float MinSeventhDown(float frequency) => frequency / 16 * 9;
-		public static float MajSeventhDown(float frequency) => frequency / 15 * 7;
-		public static float OctaveDown(float frequency) => frequency / 2;
+		public static double OctaveUp(double frequency) => frequency * 2;
+		public static double MajSeventhUp(double frequency) => frequency / 8 * 15;
+		public static double MinSeventhUp(double frequency) => frequency / 9 * 16;
+		public static double MajSixthUp(double frequency) => frequency / 3 * 5;
+		public static double MinSixthUp(double frequency) => frequency / 5 * 8;
+		public static double PerfectFifthUp(double frequency) => frequency / 2 * 3;
+		public static double TritoneUp(double frequency) => frequency / 32 * 45;
+		public static double PerfectFourthUp(double frequency) => frequency / 3 * 4;
+		public static double MajThirdUp(double frequency) => frequency / 4 * 5;
+		public static double MinThirdUp(double frequency) => frequency / 5 * 6;
+		public static double MajSecondUp(double frequency) => frequency / 8 * 9;
+		public static double MinSecondUp(double frequency) => frequency / 15 * 16;
+		public static double Unison(double frequency) => frequency;
+		public static double MinSecondDown(double frequency) => frequency / 16 * 15;
+		public static double MajSecondDown(double frequency) => frequency / 9 * 8;
+		public static double MinThirdDown(double frequency) => frequency / 6 * 5;
+		public static double MajThirdDown(double frequency) => frequency / 5 * 4;
+		public static double PerfectFourthDown(double frequency) => frequency / 4 * 3;
+		public static double TritoneDown(double frequency) => frequency / 45 * 32;
+		public static double PerfectFifthDown(double frequency) => frequency / 3 * 2;
+		public static double MinSixthDown(double frequency) => frequency / 8 * 5;
+		public static double MajSixthDown(double frequency) => frequency / 5 * 3;
+		public static double MinSeventhDown(double frequency) => frequency / 16 * 9;
+		public static double MajSeventhDown(double frequency) => frequency / 15 * 7;
+		public static double OctaveDown(double frequency) => frequency / 2;
 
 		public static Shift[] shifts = new Shift[]
 		{
